chore: sets up boilerplate to style the pdf component
