+++ src/app/components/PdfDocument/PdfDocument.module.scss
@@ -1,0 +1,10 @@
+@use '@/styles/ohma';
+
+.PdfDocument {
+    .leftPage {
+
+    }
+    .rightPage {
+
+    }
+}
